requirements.txt
python-dotenv==1.0.1
langchain==0.2.10
langchain-community==0.2.9
langchain-core==0.2.22
langchain-openai==0.1.17
langchain-text-splitters==0.2.2
openai==1.36.0
langchain_chroma==0.1.2
fastapi~=0.115.6
python-multipart
pydantic~=2.10.3
tzdata
SQLAlchemy~=2.0.36
python-json-logger~=2.0.7
starlette~=0.41.3
pydantic-settings~=2.7.0
requests~=2.32.3
Jinja2~=3.1.4
uvicorn~=0.32.1